+++ NEWS.md
@@ -2,8 +2,9 @@
 
 ## Version 2.0.0
 
-Breaking changes compared to version 1:
-- Rename `baretype` as `bare_type`.
+Breaking changes:
+- Rename `baretype` as `bare_type` and `convert_baretype` as
+  `convert_bare_type`.
 
 New features:
 - `bare_type()` with no arguments yields `Unitless.BareNumber`.
@@ -11,6 +12,8 @@ New features:
   `convert_bare_type` except for complex numbers or numeric types for which the
   bare real type backing the storage of the real and imaginary parts is
   considered.
+- `floating_point_type(args...)` yields a floating-point type appropriate to
+  represent the bare real type of `args...`.
 - `unitless(x)` yields `x` without its units if any. `x` may be a number or a
   numeric type.